Move test file to follow pattern; ensure we're testing the same createapiservice call
