+++ sdk/src/main/java/com/stytch/sdk/network/StytchApi.kt
@@ -16,17 +16,9 @@ import com.stytch.sdk.network.responseData.LoginOrCreateOTPData
 import com.stytch.sdk.network.responseData.StrengthCheckResponse
 import com.stytch.sdk.network.responseData.StytchErrorResponse
 import java.lang.RuntimeException
-import java.util.concurrent.TimeUnit
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import okhttp3.Interceptor
-import okhttp3.OkHttpClient
 import retrofit2.HttpException
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
-
-private const val ONE_HUNDRED_TWENTY = 120L
-private const val HTTP_UNAUTHORIZED = 401
 
 internal object StytchApi {
 
@@ -66,43 +58,7 @@ internal object StytchApi {
     @VisibleForTesting
     internal val apiService: StytchApiService by lazy {
         StytchClient.assertInitialized()
-        createApiService(hostUrl)
-    }
-
-    private fun createApiService(hostUrl: String): StytchApiService {
-        return Retrofit.Builder()
-            .baseUrl(hostUrl)
-            .addConverterFactory(MoshiConverterFactory.create())
-            .client(
-                OkHttpClient.Builder()
-                    .readTimeout(ONE_HUNDRED_TWENTY, TimeUnit.SECONDS)
-                    .writeTimeout(ONE_HUNDRED_TWENTY, TimeUnit.SECONDS)
-                    .connectTimeout(ONE_HUNDRED_TWENTY, TimeUnit.SECONDS)
-                    .addInterceptor(authHeaderInterceptor)
-                    .addInterceptor(
-                        Interceptor { chain ->
-                            val request = chain.request()
-                            val response = chain.proceed(request)
-                            if (response.code == HTTP_UNAUTHORIZED) {
-                                StytchClient.sessionStorage.revoke()
-                            }
-                            return@Interceptor response
-                        }
-                    )
-                    .addNetworkInterceptor {
-                        // OkHttp is adding a charset to the content-type which is rejected by the API
-                        // see: https://github.com/square/okhttp/issues/3081
-                        it.proceed(
-                            it.request()
-                                .newBuilder()
-                                .header("Content-Type", "application/json")
-                                .build()
-                        )
-                    }
-                    .build()
-            )
-            .build()
-            .create(StytchApiService::class.java)
+        StytchApiService.createApiService(hostUrl, authHeaderInterceptor)
     }
 
     internal object MagicLinks {
