+++ sdk/src/main/java/com/stytch/sdk/network/StytchApiService.kt
@@ -1,7 +1,16 @@
 package com.stytch.sdk.network
 
+import com.stytch.sdk.StytchClient
+import java.util.concurrent.TimeUnit
+import okhttp3.Interceptor
+import okhttp3.OkHttpClient
+import retrofit2.Retrofit
+import retrofit2.converter.moshi.MoshiConverterFactory
 import retrofit2.http.Body
 import retrofit2.http.POST
+
+private const val ONE_HUNDRED_TWENTY = 120L
+private const val HTTP_UNAUTHORIZED = 401
 
 @Suppress("TooManyFunctions")
 internal interface StytchApiService {
@@ -76,4 +85,44 @@ internal interface StytchApiService {
         @Body request: StytchRequests.Passwords.StrengthCheckRequest
     ): StytchResponses.Passwords.PasswordsStrengthCheckResponse
     //endregion passwords
+
+    companion object {
+        private fun clientBuilder(authHeaderInterceptor: StytchAuthHeaderInterceptor?): OkHttpClient {
+            val builder = OkHttpClient.Builder()
+                .readTimeout(ONE_HUNDRED_TWENTY, TimeUnit.SECONDS)
+                .writeTimeout(ONE_HUNDRED_TWENTY, TimeUnit.SECONDS)
+                .connectTimeout(ONE_HUNDRED_TWENTY, TimeUnit.SECONDS)
+            authHeaderInterceptor?.let { builder.addInterceptor(it) }
+            builder
+                .addInterceptor(
+                    Interceptor { chain ->
+                        val request = chain.request()
+                        val response = chain.proceed(request)
+                        if (response.code == HTTP_UNAUTHORIZED) {
+                            StytchClient.sessionStorage.revoke()
+                        }
+                        return@Interceptor response
+                    }
+                )
+                .addNetworkInterceptor {
+                    // OkHttp is adding a charset to the content-type which is rejected by the API
+                    // see: https://github.com/square/okhttp/issues/3081
+                    it.proceed(
+                        it.request()
+                            .newBuilder()
+                            .header("Content-Type", "application/json")
+                            .build()
+                    )
+                }
+            return builder.build()
+        }
+        fun createApiService(hostUrl: String, authHeaderInterceptor: StytchAuthHeaderInterceptor?): StytchApiService {
+            return Retrofit.Builder()
+                .baseUrl(hostUrl)
+                .addConverterFactory(MoshiConverterFactory.create())
+                .client(clientBuilder(authHeaderInterceptor))
+                .build()
+                .create(StytchApiService::class.java)
+        }
+    }
 }
